estudo-experimental-tdd-ct/estudo-de-caso/sem-criterios/participante01/bowling-game-microservice/src/main/java/br/edu/utfpr/bowlinggame/service/ScoreService.java
package br.edu.utfpr.bowlinggame.service;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import br.edu.utfpr.bowlinggame.dto.ScoreDTO;
import br.edu.utfpr.bowlinggame.entity.Round;
import br.edu.utfpr.bowlinggame.repository.RoundRepository;

@Service
public class ScoreService {

	@Autowired
	private RoundRepository roundRepository;
	
	public ScoreDTO getScore() {
		int sum = roundRepository.findAll().stream().mapToInt(Round::getScore).sum();
		return new ScoreDTO(sum);
	}

}
